+++ examples/multiFileProject/multiFileProject.cpp
@@ -1,5 +1,6 @@
-/******************************************************************************************************************************************
-  StoreNameAndSurname.ino
+/****************************************************************************************************************************
+  multiFileProject.cpp
+  
   For SAMD21/SAMD51 using Flash emulated-EEPROM
 
   The FlashStorage_SAMD library aims to provide a convenient way to store and retrieve user's data using the non-volatile flash memory
@@ -8,51 +9,14 @@
   Based on and modified from Cristian Maglie's FlashStorage (https://github.com/cmaglie/FlashStorage)
 
   Built by Khoi Hoang https://github.com/khoih-prog/FlashStorage_SAMD
-  Licensed under LGPLv3 license
-  
-  Orginally written by A. Christian
-  
-  Copyright (c) 2015-2016 Arduino LLC.  All right reserved.
-  Copyright (c) 2020 Khoi Hoang.
-  
-  This program is free software: you can redistribute it and/or modify it under the terms of the GNU General Public License 
-  as published bythe Free Software Foundation, either version 3 of the License, or (at your option) any later version.
-  This program is distributed in the hope that it will be useful, but WITHOUT ANY WARRANTY; without even the implied warranty of
-  MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the GNU Lesser General Public License for more details.
-  You should have received a copy of the GNU Lesser General Public License along with this library. 
-  If not, see (https://www.gnu.org/licenses/)
- ******************************************************************************************************************************************/
+*****************************************************************************************************************************/
 
-// Demonstrate how to use FlashStorage_SAMD with an API that is similar to the EEPROM library to Store and retrieve structured data.
-//#define EEPROM_EMULATION_SIZE     (4 * 1024)
+// To demo how to include files in multi-file Projects
 
-// Use 0-2. Larger for more debugging messages
-#define FLASH_DEBUG       0
+#include "multiFileProject.h"
 
-// To be included only in main(), .ino with setup() to avoid `Multiple Definitions` Linker Error
-#include <FlashStorage_SAMD.h>
-
-const int WRITTEN_SIGNATURE = 0xBEEFDEED;
-
-// Create a structure that is big enough to contain a name
-// and a surname. The "valid" variable is set to "true" once
-// the structure is filled with actual data for the first time.
-typedef struct
+void testEEPROM()
 {
-  char name[100];
-  char surname[100];
-} Person;
-
-void setup()
-{
-  Serial.begin(115200);
-  while (!Serial);
-
-  delay(200);
-
-  Serial.print(F("\nStart StoreNameAndSurname on ")); Serial.println(BOARD_NAME);
-  Serial.println(FLASH_STORAGE_SAMD_VERSION);
-
   Serial.print("EEPROM length: ");
   Serial.println(EEPROM.length());
 
@@ -113,9 +77,4 @@ void setup()
     Serial.println(" >> have been saved. Thank you!");
     Serial.println("You can reset to check emulated-EEPROM data retention.");
   }
-}
-
-void loop()
-{
-  // Do nothing...
 }