+++ src/FlashStorage_SAMD21.h
@@ -22,15 +22,19 @@
   You should have received a copy of the GNU Lesser General Public License along with this library. 
   If not, see (https://www.gnu.org/licenses/)
   
-  Version: 1.2.1
+  Version: 1.3.0
 
   Version Modified By   Date        Comments
   ------- -----------  ----------   -----------
   1.0.0   K Hoang      28/03/2020  Initial coding to add support to SAMD51 besides SAMD21
   1.1.0   K Hoang      26/01/2021  Add supports to put() and get() for writing and reading the whole object. Fix bug.
   1.2.0   K Hoang      18/08/2021  Optimize code. Add debug option
-  1.2.0   K Hoang      10/10/2021  Update `platform.ini` and `library.json`
+  1.2.1   K Hoang      10/10/2021  Update `platform.ini` and `library.json`
+  1.3.0   K Hoang      10/10/2021  Fix `multiple-definitions` linker error. Add support to many more boards.
  ******************************************************************************************************************************************/
+ 
+// The .hpp contains only definitions, and can be included as many times as necessary, without `Multiple Definitions` Linker Error
+// The .h contains implementations, and can be included only in main(), .ino with setup() to avoid `Multiple Definitions` Linker Error
 
 #pragma once