+++ changelog.md
@@ -12,6 +12,7 @@
 ## Table of Contents
 
 * [Changelog](#changelog)
+  * [Releases v1.3.0](#releases-v130)
   * [Releases v1.2.1](#releases-v121)
   * [Releases v1.2.0](#releases-v120)
   * [Releases v1.1.0](#releases-v110)
@@ -22,6 +23,14 @@
 ---
 
 ## Changelog
+
+### Releases v1.3.0
+
+1. Fix `multiple-definitions` linker error.
+2. Add support to many more boards, such as `SAMD21E1xA`, `SAMD21G1xA` and`SAMD21J1xA`
+3. Add example [multiFileProject](examples/multiFileProject) to demo for multiple-file project.
+4. Fix bug
+5. Update `Packages' Patches`
 
 ### Releases v1.2.1